fix(protocol): repair path references the rename missed
CI caught three failures, all from path strings a slash-based text rewrite
cannot see.

scripts/tests/test-runner.spec.ts builds a temp fixture tree and asserts the
runner classifies contact/tests/contact.inviter.spec.ts as a live-model spec.
LIVE_MODEL_SPECS was updated for the rename; the fixture that exercises it was
not, so the classifier stopped matching. This failed both the `check` and
`protocol-source-tests` jobs.

eval/ was never renamed, so two kinds of damage had to be undone and one had to
be applied:
- eval/canary/canary.suites.ts and eval/ops/tests/registry-corpus.spec.ts had
  their relative imports over-rewritten to ../opportunities/ and ../premises/.
- eval/viewer/tests/viewer.io.spec.ts likewise, for a baselines JSON path.
- eval/ops/tests/metadata.spec.ts builds src paths from path.join segment
  arrays -- path.join(dir, "..", "..", "..", "src", "negotiation", ...) -- with
  no slashes anywhere, so the rewrite could not match them at all. Those needed
  updating to the new directory names.

The lesson for the remaining stages: a directory rename has to account for path
strings that are not slash-delimited literals -- segment arrays, fixture trees,
and string-keyed allowlists -- and the package `test` script runs two commands
(test-runner.spec.ts then scripts/test.ts), so running only the second misses a
gate.

Verification (full package test script this time, not just its second half):
- bun run test: 2218 pass, 0 fail, 0 errors
- bun run eval:verify: all 14 suites type-checked and tested
- bun run test:scripts: 193 pass, 0 fail
- public surface: 443 exported symbols byte-identical to dev
- runtime import cycles: 0
- bun run architecture:check, check:protocol-atlas: exit 0
- services/api tsc --noEmit: clean
